refactor: Phase 2 - Create missing utility hooks and functions
Created new utilities to further reduce code duplication:

1. Created useFormValues hook (~40 lines saved)
   - Extracts form subscription pattern used in date-field and object-field
   - Provides reactive access to all form values for dynamic field behavior
   - Location: src/hooks/use-form-values.ts

2. Created getFieldInputClassName utility (~30 lines saved)
   - Standardizes error styling pattern across 10+ field components
   - Provides consistent border-destructive styling for validation errors
   - Added to src/lib/utils.ts

3. Updated field components to use new utilities:
   - date-field.tsx: Now uses useFormValues for dynamic date restrictions
   - object-field.tsx: Now uses useFormValues for dynamic text resolution
   - text-field, number-field, textarea-field, select-field: Use getFieldInputClassName
   - date-field, combobox-field, phone-field, color-picker-field: Use getFieldInputClassName
   - multicombobox-field, radio-field: Use getFieldInputClassName

4. Type consolidation analysis completed:
   - Discovered that *FieldProps and *FieldSpecificProps serve different purposes
   - *FieldProps used in config/schema definitions
   - *FieldSpecificProps used in actual components
   - Both type systems are necessary and actively used - no consolidation needed

Total reduction: ~70 lines of duplicate code eliminated
Build status: ✓ All builds passing (ESM, CJS, DTS)
