+++ packages/formedible/src/lib/utils.ts
@@ -17,3 +17,22 @@ export function normalizeOptions(
     typeof option === "string" ? { value: option, label: option } : option
   );
 }
+
+/**
+ * Utility to generate input className with error styling
+ * @param baseClassName - Base className(s) for the input
+ * @param hasErrors - Whether the field has validation errors
+ * @param additionalClassNames - Additional className(s) to merge
+ * @returns Combined className string with error styling applied
+ */
+export function getFieldInputClassName(
+  baseClassName?: string,
+  hasErrors?: boolean,
+  ...additionalClassNames: (string | undefined)[]
+): string {
+  return cn(
+    baseClassName,
+    hasErrors ? "border-destructive" : "",
+    ...additionalClassNames
+  );
+}
